% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allBivarPlots.R
\name{allBivarPlots}
\alias{allBivarPlots}
\title{Explore All Bivariate Plots}
\usage{
allBivarPlots(e)
}
\arguments{
\item{e}{(depreciated ... will be updated in future)}
}
\description{
Explore all bivariate plots
}
\details{
Opens a pop-up window that allows users to click through all bivariate plots, after
selecting the "primary variable of interest"
}
\author{
Tom Elliott
}

