% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allUniPlots.R
\name{allUniPlots}
\alias{allUniPlots}
\title{Explore All Univariate Plots}
\usage{
allUniPlots(e)
}
\arguments{
\item{e}{(depreciated ... will be updated in future)}
}
\description{
Explore all univariate plots
}
\details{
Opens a pop-up window that allows users to click through all univariate plots,
which is useful for quickly exploring large data sets.
}
\author{
Tom Elliott
}

