% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomModule.R
\docType{class}
\name{CustomModule-class}
\alias{CustomModule-class}
\alias{CustomModule}
\title{iNZight Custom Module}
\description{
iNZight Custom Module
}
\details{
Provides a basic module for extending new ones
}

\author{
Tom Elliott
}
