% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InstallModules.R
\docType{class}
\name{InstallModules-class}
\alias{InstallModules-class}
\alias{InstallModules}
\title{iNZight Module Installation}
\description{
iNZight Module Installation
}
\details{
A window to allow installation of modules from various locations.

Add-on repo -> URL -> local file
}

\author{
Daniel Barnett
}
