% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inzexpClass.R
\name{inzexp}
\alias{inzexp}
\title{iNZight Variable Object}
\usage{
inzexp(x, transform = c("none", "power", "log", "square root", "polynomial"),
  arg = 2, multiple = FALSE)
}
\arguments{
\item{x}{the name of the variable}

\item{transform}{the transformation}

\item{arg}{optional argument to the transformation}

\item{multiple}{logical, if TRUE, the variable can be used multiple times with the given transformation (e.g., power)}
}
\value{
an inzexp object
}
\description{
Default constructor for an `inzexp` object, which is a list of
variables with their allocated transformations.
}
\author{
Tom Elliott
}

