% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iNZightMaps2.R
\docType{class}
\name{iNZightMap2Mod-class}
\alias{iNZightMap2Mod-class}
\alias{iNZightMap2Mod}
\title{iNZight Maps Module 2}
\description{
iNZight Mapping Module 2
}
\details{
Opens a UI for visualising geographical data
}

\author{
Daniel Barnett
}
