% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iNZightMapsLanding.R
\docType{class}
\name{iNZightMapLanding-class}
\alias{iNZightMapLanding-class}
\alias{iNZightMapLanding}
\title{iNZight Maps Module Landing}
\description{
iNZight Mapping Module Landing Window
}
\details{
Opens a UI for visualising geographical data
}

\author{
Daniel Barnett
}
