% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iNZightMaps.R
\docType{class}
\name{iNZightMapMod-class}
\alias{iNZightMapMod-class}
\alias{iNZightMapMod}
\title{iNZight Maps Module}
\description{
iNZight Mapping Module
}
\details{
Opens a UI for visualising geographical data
}

\author{
Tom Elliott
}
