\name{iNZightModules}
\docType{package}
\title{
  Suite of Add-on Modules for iNZight GUI software.
}
\description{
  This package contains a suite of add-on GUI modules for the iNZight
  program. It is no intendened for standalone use. Major modules include
  Model Fitting and Time Series.
}
\details{
\tabular{ll}{
  Package: \tab iNZightTools\cr
  Type: \tab Package\cr
  Version: \tab 1.0\cr
  Date: \tab 2014-02-25\cr
  License: \tab GPL-2\cr
}
}
\author{
  Junjie Zeng. Original work by Simon Potter, Dineika Chandranada, and
  David Banks. 
  
  Maintainer: Junjie Zeng <jzen188@aucklanduni.ac.nz>
}
\keyword{ iNZight }
\seealso{
  \code{\link{iNZight}}, \code{\link{iNZightTools}}
}
\examples{

}
