% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iNZightModules-package.R
\docType{package}
\name{iNZightModules}
\alias{iNZightModules}
\alias{iNZightModules-package}
\title{iNZightModules: Add-on modules for iNZight}
\description{
Each file in this package is a self-contained module that integrates into iNZight.
Each module targets a specific R package (for example, iNZightTS, iNZightMaps, etc).
}
\author{
Tom Elliott, Eric Lim, Junjie Zeng, Simon Potter, Dineika Chandranada, David Banks

Maintainer: Tom Elliott <inzight_support@stat.auckland.ac.nz>
}

