% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iNZightRegression.R
\docType{class}
\name{iNZightRegMod-class}
\alias{iNZightRegMod}
\alias{iNZightRegMod-class}
\title{iNZight Model Fitting Module}
\description{
iNZight Model Fitting Module
}
\details{
A GUI add-on for fitting regression models.
}
\author{
Tom Elliott
}

