% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iNZightTSLegacy.R
\docType{class}
\name{iNZightTSLegacyMod-class}
\alias{iNZightTSLegacyMod-class}
\alias{iNZightTSLegacyMod}
\title{iNZight Time Series Module}
\description{
iNZight Time Series Module
}
\details{
A GUI add-on for visualising and doing basic inference and prediction of time series data.
}

\author{
Eric Lim
}
