% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelFitting.R
\name{modelFitting}
\alias{modelFitting}
\title{Model Fitting Module}
\usage{
modelFitting(e)
}
\arguments{
\item{e}{going out of fashion... needs to be written as a reference class.}
}
\description{
A GUI for fitting ([survey] generalised) linear models.
}
\details{
Provides a UI - probably needs rewriting (using RC) to avoid stuff like `<<-` and
attaching to the Global Environment.
}
\author{
Tom Elliott
}
