% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModuleManager.R
\docType{class}
\name{ModuleManager-class}
\alias{ModuleManager-class}
\alias{ModuleManager}
\title{iNZight Module Installation}
\description{
iNZight Module Installation
}
\details{
A window to allow installation of modules from various locations.
}

\author{
Tom Elliott, Daniel Barnett
}
