% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot3Dmode.R
\name{plot3D}
\alias{plot3D}
\title{3D Plot}
\usage{
plot3D(e)
}
\arguments{
\item{e}{depreciated ... will be removed}
}
\description{
Plot 3D: Three-dimentional, interactive graphics
}
\details{
Allows users to plot three continuous variables, as well as one categorical variable.
}

