% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportPlot3D.R
\name{scatter3d}
\alias{scatter3d}
\title{Plot three variable in a 3D interactive graphics window}
\usage{
scatter3d(x, y, z, xlab = deparse(substitute(x)),
  ylab = deparse(substitute(y)), axis.scales = TRUE,
  zlab = deparse(substitute(z)), revolutions = 0, bg.col = c("white",
  "black"), axis.col = if (bg.col == "white") c("darkmagenta", "black",
  "darkcyan") else c("darkmagenta", "white", "darkcyan"),
  surface.col = c("blue", "green", "orange", "magenta", "cyan", "red",
  "yellow", "gray"), neg.res.col = "red", pos.res.col = "green",
  square.col = if (bg.col == "white") "black" else "gray",
  point.col = "yellow", text.col = axis.col, grid.col = if (bg.col ==
  "white") "black" else "gray", fogtype = c("exp2", "linear", "exp", "none"),
  residuals = (length(fit) == 1), surface = TRUE, fill = TRUE,
  grid = TRUE, grid.lines = 26, df.smooth = NULL, df.additive = NULL,
  sphere.size = 1, threshold = 0.01, speed = 1, fov = 60,
  fit = "linear", groups = NULL, parallel = TRUE, ellipsoid = FALSE,
  level = 0.5, model.summary = FALSE)
}
\arguments{
\item{x}{a parameter}

\item{y}{a parameter}

\item{z}{a parameter}

\item{xlab}{a parameter}

\item{ylab}{a parameter}

\item{axis.scales}{a parameter}

\item{zlab}{a parameter}

\item{revolutions}{a parameter}

\item{bg.col}{a parameter}

\item{axis.col}{a parameter}

\item{surface.col}{a parameter}

\item{neg.res.col}{a parameter}

\item{pos.res.col}{a parameter}

\item{square.col}{a parameter}

\item{point.col}{a parameter}

\item{text.col}{a parameter}

\item{grid.col}{a parameter}

\item{fogtype}{a parameter}

\item{residuals}{a parameter}

\item{surface}{a parameter}

\item{fill}{a parameter}

\item{grid}{a parameter}

\item{grid.lines}{a parameter}

\item{df.smooth}{a parameter}

\item{df.additive}{a parameter}

\item{sphere.size}{a parameter}

\item{threshold}{a parameter}

\item{speed}{a parameter}

\item{fov}{a parameter}

\item{fit}{a parameter}

\item{groups}{a parameter}

\item{parallel}{a parameter}

\item{ellipsoid}{a parameter}

\item{level}{a parameter}

\item{model.summary}{a parameter}
}
\description{
Description ...
}
\details{
Details ...
}

