% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterPlotMatrix.R
\name{scatterPlotMatrix}
\alias{scatterPlotMatrix}
\title{Scatterplot Matrix}
\usage{
scatterPlotMatrix(e, ...)
}
\arguments{
\item{e}{going, going, going ...}

\item{...}{more opts}
}
\description{
Draws a matrix of all plot-combinations based on selected variables.
}
\details{
User selects variables, they are drawn.
}
\author{
iNZight Team
}

