% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomModule.R
\docType{class}
\name{InstallModules-class}
\alias{InstallModules-class}
\alias{InstallModules}
\title{iNZight Module Installation}
\description{
iNZight Module Installation
}
\details{
A window to allow installation/removal of iNZight addon modules
}

\author{
Daniel Barnett
}
